Added new routes, views and updated the store
Added the podcast route to view details about a specific podcast. Updated the scroll behaviour to default to the top on every route access. Added the Podcast view to display details about a specific podcast. Added the searchString, podcastDetails and podcastEpisodes variables to the state and also added their corresponding getters. Added the setSearchString mutation to update the searchString state variable. Added the SET_PODCAST_DETAILS and SET_EPISODE_DETAILS mutations to update podcast and episode details. Added the fetchPodcastsByFeedId and fetchEpisodesByFeedId actions to fetch podcast and episode details. Removed the PageLoader component and replaced it with the Loader component. Updated image size and text overflow in the PodcastCard component. Updated card display in the PodcastCardGrid component. Added the PodcastEpisodeCard component to display episode details of a podcast. Added the ContentCollapse component that to display more or less data. Updated the search bar rendering logic on the home page.