Added font awesome and sass packages
Added the fontawesome-svg-core, free-regular-svg-icons, free-solid-svg-icons and vue-fontawesome packages. Added sass and sass-loader packages. Added the vue-config file and configuration for scss. Added the base.scss with modifications of the base buefy theme. Updated main.js with icons to be used in the project. Added lobster and robot fonts CDN in the index.html.